pyproject.toml
[project]
name = "create-contract"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docx2pdf>=0.1.8",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "python-docx>=1.1.2",
    "ruff>=0.7.1",
]
[tool.ruff]
extend-select = ["C4", "SIM", "TCH"]
ignore = []
show-fixes = true
target-version = "py37"

[tool.uv.workspace]
members = ["ex"]
